requirements.txt
aiocontextvars==0.2.1
aiohttp==3.5.4
async-timeout==3.0.1
asyncpg==0.18.2
attrs==18.2.0
chardet==3.0.4
contextvars==2.3
gino==0.8.1
idna==2.8
idna-ssl==1.1.0
immutables==0.6
multidict==4.5.2
SQLAlchemy==1.2.16
typing-extensions==3.7.2
yarl==1.3.0
